Updated Solution - GfG to GitHub

+++ Difficulty: Medium/nCr/n-cr.cpp
@@ -7,12 +7,9 @@ using namespace std;
 
 // } Driver Code Ends
 
-// User function Template for C++
-
 class Solution {
   public:
-    long long nCr(int n, int r) 
-    {
+    int nCr(int n, int r) {
         if (r > n) return 0;
         if (r == 0 || r == n) return 1;
         long long res = 1;
